requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.3
django-on-heroku==1.1.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.0.4
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
mirakuru==2.4.2
numpy==1.23.4
openai==0.25.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pandas-stubs==1.5.1.221024
pluggy==1.0.0
port-for==0.6.2
psutil==5.9.4
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytest==7.2.0
pytest-postgresql==4.1.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.12
whitenoise==6.2.0
